Extract getLastItem
